refactor: 카드에 추가 함수로 수정
